Moves LoadingSpinner, OwnerEditor, PopularTables, Preloader, SearchBar
