+++ amundsen_application/static/js/components/PopularTables/index.tsx
