+++ amundsen_application/static/js/components/PopularTables/styles.scss
