+++ amundsen_application/static/js/features/Feedback/FeedbackForm/index.spec.tsx
@@ -5,9 +5,9 @@ import * as React from 'react';
 
 import { shallow } from 'enzyme';
 
-import LoadingSpinner from 'components/common/LoadingSpinner';
+import LoadingSpinner from 'components/LoadingSpinner';
 import { SendingState } from 'interfaces';
-import FeedbackForm, { FeedbackFormProps } from '.';
+import { FeedbackFormProps } from '.';
 import { RatingFeedbackForm } from './RatingFeedbackForm';
 
 import { SUBMIT_FAILURE_MESSAGE, SUBMIT_SUCCESS_MESSAGE } from '../constants';
